Tune output - writes to a file
For the first time, I think the formula is at a state where it produces
believable results. Still plenty of bugs and tuning to work out though.

One observation is the distribution of OB values - I noticed a 3, plenty
of 4's and a lot of high numbers as well, 12 even, lots of 10s. Maybe
tighten it up just a bit, unless there are convincing reasons these are
more correct.